+++ handlers/product-move-api/src/test/scala/com/gu/productmove/zuora/MockGetInvoice.scala
@@ -2,7 +2,7 @@ package com.gu.productmove.zuora
 
 import com.gu.productmove.endpoint.move.ProductMoveEndpointTypes.{ErrorResponse, InternalServerError}
 import com.gu.productmove.zuora.GetInvoice.GetInvoiceResponse
-import com.gu.productmove.zuora.RunBilling.InvoiceId
+import com.gu.productmove.zuora.model.InvoiceId
 import zio.*
 
 class MockGetInvoice(
