+++ handlers/product-move-api/src/test/scala/com/gu/productmove/zuora/MockTermRenewal.scala
@@ -5,7 +5,7 @@ import com.gu.newproduct.api.productcatalog.BillingPeriod
 import com.gu.productmove.GuStageLive
 import com.gu.productmove.endpoint.move.ProductMoveEndpointTypes.{ErrorResponse, InternalServerError, PreviewResult}
 import com.gu.productmove.zuora.GetSubscription.GetSubscriptionResponse
-import com.gu.productmove.zuora.RunBilling.InvoiceId
+import com.gu.productmove.zuora.model.InvoiceId
 import com.gu.productmove.zuora.{GetSubscription, SubscriptionUpdatePreviewResponse}
 import com.gu.productmove.zuora.model.SubscriptionName
 import zio.json.JsonDecoder
