+++ handlers/product-move-api/src/test/scala/com/gu/productmove/zuora/manual/CreatePaymentSpec.scala
@@ -4,7 +4,7 @@ import com.gu.productmove.*
 import com.gu.productmove.endpoint.move.ProductMoveEndpoint
 import com.gu.productmove.endpoint.move.ProductMoveEndpointTypes.ExpectedInput
 import com.gu.productmove.endpoint.move.switchtype.RecurringContributionToSupporterPlus
-import com.gu.productmove.zuora.RunBilling.InvoiceId
+import com.gu.productmove.zuora.model.InvoiceId
 import com.gu.productmove.zuora.model.SubscriptionName
 import com.gu.productmove.zuora.rest.{ZuoraClientLive, ZuoraGetLive}
 import com.gu.productmove.zuora.{CreatePayment, CreatePaymentLive, ZuoraAccountId}
