+++ handlers/product-move-api/src/test/scala/com/gu/productmove/zuora/manual/PostInvoicesSpec.scala
@@ -1,7 +1,7 @@
 package com.gu.productmove.zuora.manual
 
 import com.gu.productmove.GuStageLive.Stage
-import com.gu.productmove.zuora.RunBilling.InvoiceId
+import com.gu.productmove.zuora.model.InvoiceId
 import com.gu.productmove.zuora.{PostInvoicesLive, TermRenewalLive}
 import com.gu.productmove.zuora.model.SubscriptionName
 import com.gu.productmove.zuora.rest.{ZuoraClientLive, ZuoraGet, ZuoraGetLive}
